chore: lower linecheck gate to 700 lines
Split nine oversized source files (including production modules cli.rs
and service.rs) into sibling modules so every .rs file stays under 700
lines. Updates .githooks/pre-push to enforce the new limit.
