+++ .changeset/linecheck-700.md
@@ -1,0 +1,5 @@
+---
+"moadim": patch
+---
+
+chore: lower linecheck gate from 1000 to 700 lines
