+++ src/routes/http_tests.rs
@@ -3,40 +3,10 @@
 use axum::{
     body::Body,
     http::{header::CONTENT_TYPE, Request, StatusCode},
-    routing::post,
-    Router,
 };
 use tower::ServiceExt;
 
-use super::{build_app, echo, write_openapi_spec};
-
-/// Point `MOADIM_HOME_OVERRIDE` at a fresh, empty temp home for the duration of a test, removing it
-/// on drop. With no agent TOMLs present, agent validation falls back to the built-in names (so
-/// `"claude"` is accepted) while `load_agent_command` finds no config — exercising the trigger
-/// "no spawn" path without launching a real agent or writing into the user's real home. Tests in
-/// this crate run single-threaded per binary, so the global env mutation is safe.
-struct TempHome;
-
-impl TempHome {
-    fn set() -> Self {
-        let dir = std::env::temp_dir().join(format!("moadim-httptest-{}", uuid::Uuid::new_v4()));
-        std::fs::create_dir_all(&dir).expect("create temp home");
-        // SAFETY: single-threaded test execution.
-        unsafe {
-            std::env::set_var("MOADIM_HOME_OVERRIDE", &dir);
-        }
-        Self
-    }
-}
-
-impl Drop for TempHome {
-    fn drop(&mut self) {
-        // SAFETY: single-threaded test execution.
-        unsafe {
-            std::env::remove_var("MOADIM_HOME_OVERRIDE");
-        }
-    }
-}
+use super::{build_app, write_openapi_spec};
 
 // ── openapi spec writer ──────────────────────────────────────────────────────
 
@@ -507,389 +477,6 @@ async fn router_unknown_api_path_non_get_returns_404() {
                 .method("POST")
                 .uri("/api/v1/bogus")
                 .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
-}
-
-#[tokio::test]
-async fn router_routines_cleanup_returns_removed_count() {
-    let resp = build_app(crate::routines::new_store())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/v1/routines/cleanup")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-    let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let val: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-    assert!(val["removed"].is_u64());
-}
-
-// ── echo handler ──────────────────────────────────────────────────────────────
-
-#[tokio::test]
-async fn echo_returns_message_and_timestamp() {
-    let app = Router::new().route("/echo", post(echo));
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/echo")
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(r#"{"message":"hello"}"#))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-    let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let json: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-    assert_eq!(json["message"], "hello");
-    assert!(json["timestamp"].as_u64().is_some());
-}
-
-#[tokio::test]
-async fn echo_rejects_invalid_json() {
-    let app = Router::new().route("/echo", post(echo));
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/echo")
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from("not-json"))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-}
-
-#[tokio::test]
-async fn echo_rejects_missing_message_field() {
-    let app = Router::new().route("/echo", post(echo));
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/echo")
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(r#"{"other":"field"}"#))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-}
-
-// ── routines CRUD lifecycle (covers all routine HTTP handlers) ────────────────
-
-#[tokio::test]
-async fn router_routine_full_lifecycle() {
-    let _home = TempHome::set();
-    let routines = crate::routines::new_store();
-
-    let body = r#"{"schedule":"@daily","title":"Http Routine","agent":"claude","prompt":"p","repositories":[{"repository":"r","branch":"main"}]}"#;
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/v1/routines")
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(body))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::CREATED);
-    let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let created: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-    let id = created["id"].as_str().unwrap().to_string();
-
-    // GET list
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .uri("/api/v1/routines")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-
-    // GET one
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .uri(format!("/api/v1/routines/{id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-
-    // PATCH
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("PATCH")
-                .uri(format!("/api/v1/routines/{id}"))
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(r#"{"title":"Patched"}"#))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-
-    // PUT (replace)
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("PUT")
-                .uri(format!("/api/v1/routines/{id}"))
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(r#"{"prompt":"replaced"}"#))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-
-    // trigger (records the manual trigger and returns OK)
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri(format!("/api/v1/routines/{id}/trigger"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-
-    // scheduled-trigger (the crontab-invoked path; runs the routine and returns OK)
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri(format!("/api/v1/routines/{id}/scheduled-trigger"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-
-    // logs (empty)
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .uri(format!("/api/v1/routines/{id}/logs"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-
-    // POST flag
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri(format!("/api/v1/routines/{id}/flags"))
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(
-                    r#"{"type":"bug","description":"broken thing","scope":"general"}"#,
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::CREATED);
-    let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let flag: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-    let filename = flag["filename"].as_str().unwrap().to_string();
-
-    // GET flags
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .uri(format!("/api/v1/routines/{id}/flags"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-    let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let flags: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-    assert_eq!(flags.as_array().unwrap().len(), 1);
-
-    // DELETE flag (resolve)
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("DELETE")
-                .uri(format!("/api/v1/routines/{id}/flags/{filename}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::NO_CONTENT);
-
-    // DELETE
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("DELETE")
-                .uri(format!("/api/v1/routines/{id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-    assert!(!crate::paths::routine_dir(&id).exists());
-}
-
-#[tokio::test]
-async fn router_flag_create_rejects_bad_scope() {
-    let _home = TempHome::set();
-    let routines = crate::routines::new_store();
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/v1/routines")
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(
-                    r#"{"schedule":"@daily","title":"Flag Scope Routine","agent":"claude","prompt":"p"}"#,
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let id = serde_json::from_slice::<serde_json::Value>(&bytes).unwrap()["id"]
-        .as_str()
-        .unwrap()
-        .to_string();
-
-    let resp = build_app(routines.clone())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri(format!("/api/v1/routines/{id}/flags"))
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(
-                    r#"{"type":"bug","description":"d","scope":"nowhere"}"#,
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-}
-
-#[tokio::test]
-async fn router_flag_not_found_paths() {
-    let resp = build_app(crate::routines::new_store())
-        .oneshot(
-            Request::builder()
-                .uri("/api/v1/routines/no-such/flags")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
-
-    let resp = build_app(crate::routines::new_store())
-        .oneshot(
-            Request::builder()
-                .method("DELETE")
-                .uri("/api/v1/routines/no-such/flags/bug-1.md")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::NOT_FOUND);
-}
-
-#[tokio::test]
-async fn router_routine_create_invalid_cron_400() {
-    let resp = build_app(crate::routines::new_store())
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/v1/routines")
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(
-                    r#"{"schedule":"bad","title":"t","agent":"a","prompt":"p"}"#,
-                ))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-}
-
-#[tokio::test]
-async fn router_routine_not_found_paths() {
-    for (method, suffix) in [
-        ("GET", ""),
-        ("DELETE", ""),
-        ("POST", "/trigger"),
-        ("POST", "/scheduled-trigger"),
-        ("GET", "/logs"),
-    ] {
-        let resp = build_app(crate::routines::new_store())
-            .oneshot(
-                Request::builder()
-                    .method(method)
-                    .uri(format!("/api/v1/routines/no-such{suffix}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(resp.status(), StatusCode::NOT_FOUND, "{method} {suffix}");
-    }
-
-    // PATCH nonexistent
-    let resp = build_app(crate::routines::new_store())
-        .oneshot(
-            Request::builder()
-                .method("PATCH")
-                .uri("/api/v1/routines/no-such")
-                .header(CONTENT_TYPE, "application/json")
-                .body(Body::from(r#"{"title":"x"}"#))
                 .unwrap(),
         )
         .await
