+++ src/routes/mcp.rs
@@ -486,3 +486,7 @@ impl MoadimMcp {
 #[cfg(test)]
 #[path = "mcp_tests.rs"]
 mod mcp_tests;
+
+#[cfg(test)]
+#[path = "mcp_lock_tests.rs"]
+mod mcp_lock_tests;
