+++ src/routines/mod.rs
@@ -38,3 +38,7 @@ pub(crate) use command::*;
 #[cfg(test)]
 #[path = "mod_tests.rs"]
 mod routines_tests;
+
+#[cfg(test)]
+#[path = "mod_agents_tests.rs"]
+mod mod_agents_tests;
